Added injections
Added injections. This is an unmapped and undocumented function which is not yet used in the current public release builds of Vortex Digital Control. Please do not add this to your lang folder, whilst it won't add any further functionality, it might induce instability within your Vortex Software. However, if you want to unlock beta features in Vortex Performance, add this to the folder titled assets.752.assets folder.